Pull out the shared parts of the assembly loaders
`SimpleAnalyzerAssemblyLoader` and `ShadowCopyAssemblyLoader` have a lot
of similar code. This commit pulls the shared parts out into
`AbstractAnalyzerAssemblyLoader`.
