+++ src/Compilers/Core/Desktop/ShadowCopyAnalyzerAssemblyLoader.cs
@@ -1,26 +1,15 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.IO;
-using System.Linq;
 using System.Reflection;
-using System.Reflection.Metadata;
-using System.Reflection.PortableExecutable;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Microsoft.CodeAnalysis
 {
-    public sealed class ShadowCopyAnalyzerAssemblyLoader : IAnalyzerAssemblyLoader
+    public sealed class ShadowCopyAnalyzerAssemblyLoader : AbstractAnalyzerAssemblyLoader
     {
-        private readonly List<string> _dependencyPaths = new List<string>();
-        private readonly Dictionary<string, Assembly> _pathsToAssemblies = new Dictionary<string, Assembly>(StringComparer.OrdinalIgnoreCase);
-        private readonly Dictionary<string, Assembly> _namesToAssemblies = new Dictionary<string, Assembly>();
-        private readonly object _guard = new object();
-
         /// <summary>
         /// The base directory for shadow copies. Each instance of
         /// <see cref="ShadowCopyAnalyzerAssemblyLoader"/> gets its own
@@ -40,8 +29,6 @@ namespace Microsoft.CodeAnalysis
         /// Used to generate unique names for per-assembly directories.
         /// </summary>
         private int _assemblyDirectoryId = 0;
-
-        private bool _hookedAssemblyResolve = false;
 
         public ShadowCopyAnalyzerAssemblyLoader(string baseDirectory = null)
         {
@@ -88,49 +75,7 @@ namespace Microsoft.CodeAnalysis
             }
         }
 
-        public void AddDependencyLocation(string fullPath)
-        {
-            lock (_guard)
-            {
-                if (fullPath == null)
-                {
-                    throw new ArgumentNullException(nameof(fullPath));
-                }
-
-                lock (_guard)
-                {
-                    if (!_dependencyPaths.Contains(fullPath, StringComparer.OrdinalIgnoreCase))
-                    {
-                        _dependencyPaths.Add(fullPath);
-                    }
-                }
-            }
-        }
-
-        public Assembly LoadFromPath(string fullPath)
-        {
-            lock (_guard)
-            {
-                Assembly assembly;
-                if (_pathsToAssemblies.TryGetValue(fullPath, out assembly))
-                {
-                    return assembly;
-                }
-
-                assembly = LoadCore(fullPath);
-
-                if (!_hookedAssemblyResolve)
-                {
-                    _hookedAssemblyResolve = true;
-
-                    AppDomain.CurrentDomain.AssemblyResolve += CurrentDomain_AssemblyResolve;
-                }
-
-                return assembly;
-            }
-        }
-
-        private Assembly LoadCore(string fullPath)
+        protected override Assembly LoadCore(string fullPath)
         {
             if (_shadowCopyDirectory == null)
             {
@@ -140,13 +85,7 @@ namespace Microsoft.CodeAnalysis
             string assemblyDirectory = CreateUniqueDirectoryForAssembly();
             string shadowCopyPath = CopyFileAndResources(fullPath, assemblyDirectory);
 
-            Assembly assembly = Assembly.LoadFrom(shadowCopyPath);
-            string assemblyName = assembly.FullName;
-
-            _pathsToAssemblies[fullPath] = assembly;
-            _namesToAssemblies[assemblyName] = assembly;
-
-            return assembly;
+            return Assembly.LoadFrom(shadowCopyPath);
         }
 
         private string CopyFileAndResources(string fullPath, string assemblyDirectory)
@@ -238,90 +177,6 @@ namespace Microsoft.CodeAnalysis
             Directory.CreateDirectory(directory);
 
             return directory;
-        }
-
-        private Assembly CurrentDomain_AssemblyResolve(object sender, ResolveEventArgs args)
-        {
-            string requestedNameWithPolicyApplied = AppDomain.CurrentDomain.ApplyPolicy(args.Name);
-
-            lock (_guard)
-            {
-                Assembly assembly;
-                if (_namesToAssemblies.TryGetValue(requestedNameWithPolicyApplied, out assembly))
-                {
-                    return assembly;
-                }
-
-                AssemblyIdentity requestedAssemblyIdentity;
-                if (!AssemblyIdentity.TryParseDisplayName(requestedNameWithPolicyApplied, out requestedAssemblyIdentity))
-                {
-                    return null;
-                }
-
-                foreach (string candidatePath in _dependencyPaths)
-                {
-                    if (AssemblyAlreadyLoaded(candidatePath) ||
-                        !FileMatchesAssemblyName(candidatePath, requestedAssemblyIdentity.Name))
-                    {
-                        continue;
-                    }
-
-                    AssemblyIdentity candidateIdentity = TryGetAssemblyIdentity(candidatePath);
-
-                    if (requestedAssemblyIdentity.Equals(candidateIdentity))
-                    {
-                        return LoadCore(candidatePath);
-                    }
-                }
-
-                return null;
-            }
-        }
-
-        private bool AssemblyAlreadyLoaded(string path)
-        {
-            return _pathsToAssemblies.ContainsKey(path);
-        }
-
-        private bool FileMatchesAssemblyName(string path, string assemblySimpleName)
-        {
-            return Path.GetFileNameWithoutExtension(path).Equals(assemblySimpleName, StringComparison.OrdinalIgnoreCase);
-        }
-
-        private static AssemblyIdentity TryGetAssemblyIdentity(string filePath)
-        {
-            try
-            {
-                if (!File.Exists(filePath))
-                {
-                    return null;
-                }
-
-                using (var stream = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.Read | FileShare.Delete))
-                using (var peReader = new PEReader(stream))
-                {
-                    var metadataReader = peReader.GetMetadataReader();
-
-                    AssemblyDefinition assemblyDefinition = metadataReader.GetAssemblyDefinition();
-
-                    string name = metadataReader.GetString(assemblyDefinition.Name);
-                    Version version = assemblyDefinition.Version;
-
-                    StringHandle cultureHandle = assemblyDefinition.Culture;
-                    string cultureName = (!cultureHandle.IsNil) ? metadataReader.GetString(cultureHandle) : null;
-                    AssemblyFlags flags = assemblyDefinition.Flags;
-
-                    bool hasPublicKey = (flags & AssemblyFlags.PublicKey) != 0;
-                    BlobHandle publicKeyHandle = assemblyDefinition.PublicKey;
-                    ImmutableArray<byte> publicKeyOrToken = !publicKeyHandle.IsNil
-                        ? metadataReader.GetBlobBytes(publicKeyHandle).AsImmutableOrNull()
-                        : default(ImmutableArray<byte>);
-                    return new AssemblyIdentity(name, version, cultureName, publicKeyOrToken, hasPublicKey);
-                }
-            }
-            catch { }
-
-            return null;
         }
     }
 }
